backend/requirements.txt
django
python-dotenv
djangorestframework
django-cors-headers 
